ChatterboxServer/requirements.txt
# Chatterbox TTS Server Dependencies
chatterbox-tts>=0.1.6
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
numpy>=1.24.0
soundfile>=0.12.0
